+++ packages/ckeditor5-build-balloon/tests/manual/ckeditor.js
@@ -3,9 +3,9 @@
  * For licensing, see LICENSE.md.
  */
 
-import BalloonToolbarEditor from '../../build/ckeditor';
+import BalloonEditor from '../../build/ckeditor';
 
-BalloonToolbarEditor.create( document.querySelector( '#editor' ) )
+BalloonEditor.create( document.querySelector( '#editor' ) )
 	.then( editor => {
 		window.editor = editor;
 	} )