+++ packages/ckeditor5-build-balloon/CHANGELOG.md
@@ -5,4 +5,4 @@ Changelog
 
 ### Features
 
-* Introduced the editor with a balloon toolbar build. ([82fe8d5](https://github.com/ckeditor/ckeditor5-build-balloon-toolbar/commit/82fe8d5))
+* Introduced the balloon editor build. ([82fe8d5](https://github.com/ckeditor/ckeditor5-build-balloon/commit/82fe8d5))